MORPH-12625: Address PR review comments on whitelabel test
- Move the PNG fixture helper to testhelpers.WritePNG
- Move the images test into resource_test.go and delete resource_images_test.go
- Drop enabled/appliance_name and the doc comment from the images test
- Simplify the deferred close to f.Close()
- Make enabled Optional+Computed (UseStateForUnknown) so omitting it no longer
  yields an inconsistent-result error; guard it with IsUnknown in the request builder
